Started working on the main story
Alright small rant time

So right now i am starting to work out the story port, and i made it where it will count up to what point it is in the song for me to time it everything. i also made a function to change the image. i don't have any sprites for that because i am redrawing everything. i am ending it off here because i want to get some sleep. i will keep working on this in the morning...